[+FEATURE] Soap: Handle exceptions with SOAP faults
Adds exception handling to the ServiceWrapper and converts exceptions.
If an exception was declared as "@throws" in the service method it will
be exposed to the client as a client type (faultcode) exception.
Otherwise settings can be used to log exceptions with additional data
and not to expose information to the client in production context or
directly output a SOAP fault with extended information in development
context.

Change-Id: Id26befffbb10cb17e332bcf65764b50b66efd776